Resolve module-diff from filesystem
Currently the module diff is calculated by comparing the new module lock
to the old module lock.

This isn't good enough as it doesn't account for cases where:

1) The modules on disk are removed or altered independently from kl and
   things are now out-of-sync.
2) The lockfile itself is manually modified (perhaps to rollback).

In these cases we need to compare what is on disk with the current
module lock and use that diff to download/remove modules.